+++ src/extensionContext.ts
@@ -8,6 +8,7 @@ export class ExtensionContext {
   public readonly configuration: Configuration;
   public readonly exceptionManager: ExceptionManager;
   public readonly log: vscode.LogOutputChannel;
+  private readonly pendingRestart: Map<string, boolean>;
   public readonly rubyEnvProvider: RubyEnvProvider;
   public readonly skipPathProvider: SkipPathProvider;
 
@@ -16,6 +17,7 @@ export class ExtensionContext {
     this.log = vscode.window.createOutputChannel('Tracciatto', { log: true });
 
     this.exceptionManager = new ExceptionManager(this.extensionContext, this.log);
+    this.pendingRestart = new Map();
     this.rubyEnvProvider = new RubyEnvProvider(this.configuration, this.log);
     this.skipPathProvider = new SkipPathProvider(this.configuration, this.log);
     this.disposables.push(this.exceptionManager, this.rubyEnvProvider, this.log);
@@ -23,5 +25,16 @@ export class ExtensionContext {
 
   public get disposables(): vscode.Disposable[] {
     return this.extensionContext.subscriptions;
+  }
+
+  public resetPendingRestart(sessionId: string): boolean {
+    const removed = this.pendingRestart.delete(sessionId);
+    this.log.trace(`[${sessionId.slice(0, 8)}]: Reset pending restart. WasSet: ${removed}`);
+    return removed;
+  }
+
+  public setPendingRestart(sessionId: string): void {
+    this.log.trace(`[${sessionId.slice(0, 8)}]: Set pending restart`);
+    this.pendingRestart.set(sessionId, true);
   }
 }